User can run a VM with parenthesis in its name.
When user put a parenthesis into a VM name in CernVM-Online, libcernvm
did not allowed to start it, due to Bash unsafe characters (i.e.
parenthesis). Right now, when such a request comes, we silently
erase the unsafe characters from the name (when we are starting the
machine).

For further details see: https://sft.its.cern.ch/jira/browse/CVM-1056